% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R
\name{bracket}
\alias{bracket}
\title{Build a single-elimination tournament bracket for `n` ranked teams}
\usage{
bracket(teams = NULL, n = length(teams))
}
\arguments{
\item{teams}{a list of teams, from best to worst (character/vector)}

\item{n}{number of teams (numeric/scalar)}
}
\value{
a tibble with tournament schedule: {game, round, away, home}
}
\description{
Build a single-elimination tournament bracket for `n` ranked teams
}
