% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_robin.R
\name{random_robin}
\alias{random_robin}
\title{Build a random-robin tournament}
\usage{
random_robin(teams = NULL, n = length(teams), n_games = choose(n, 2))
}
\arguments{
\item{teams}{a list of unranked teams (character/vector)}

\item{n}{number of teams (numeric/scalar)}

\item{n_games}{how many total tournament games? (numeric/scalar)}
}
\value{
a tibble with tournament schedule: {game, away, home}
}
\description{
Build a random-robin tournament
}
