% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_robin.R
\name{round_robin}
\alias{round_robin}
\title{Build a round-robin tournament}
\usage{
round_robin(teams = NULL, n = length(teams), home_away = FALSE)
}
\arguments{
\item{teams}{a list of teams, from best to worst (character/vector)}

\item{n}{number of teams (numeric/scalar)}

\item{home_away}{should each team play twice? (1 home, 1 away) (logical/scalar)}
}
\value{
a tibble with tournament schedule: {game, away, home}
}
\description{
Build a round-robin tournament
}
