% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tournament.R
\docType{data}
\name{Tournament}
\alias{Tournament}
\title{Tournament R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Tournament
}
\description{
Tournament R6 Class
}
\keyword{datasets}
